Move story saving methods out of PageContainer.cs
Changed to improve readability. Story saving methods are now kept within the SaveStoryHandler script and the open and close menu scripts are now stored within the SaveStoryButton script.
