+++ Assets/Scenes/Sentence Builder/SentenceBuilder.unity
@@ -1678,8 +1678,8 @@ MonoBehaviour:
   m_OnEndEdit:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 526344200}
-        m_TargetAssemblyTypeName: PageContainer, Assembly-CSharp
+      - m_Target: {fileID: 2100211433}
+        m_TargetAssemblyTypeName: SaveStoryButton, Assembly-CSharp
         m_MethodName: CloseStoryNameMenu
         m_Mode: 1
         m_Arguments:
@@ -1780,6 +1780,55 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 268673459}
   m_CullTransparentMesh: 0
+--- !u!1 &301490247
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 301490248}
+  - component: {fileID: 301490249}
+  m_Layer: 5
+  m_Name: SaveStoryHandler
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &301490248
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 301490247}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 1122934824}
+  m_RootOrder: 10
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 663, y: -24}
+  m_SizeDelta: {x: 100, y: 100}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &301490249
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 301490247}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac0472cd64bf8d745a7216b4299ad3c8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  PageContainer: {fileID: 526344197}
 --- !u!1 &346707162
 GameObject:
   m_ObjectHideFlags: 0
@@ -2064,7 +2113,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -142.99988, y: 252.00023}
+  m_AnchoredPosition: {x: -142.99988, y: 252.0002}
   m_SizeDelta: {x: 280, y: 500}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &396208986
@@ -3121,8 +3170,6 @@ MonoBehaviour:
   pagePrefabNoPicture: {fileID: 6817511528840321722, guid: 35a67f501cf6d6d488fd2fe1609610d1,
     type: 3}
   pageIconContainer: {fileID: 1145678817}
-  StoryNamePrompt: {fileID: 1254632426}
-  StorySubmissionStatus: {fileID: 1462654760}
   maxPageCount: 30
   selectedPageGlobal: -1
 --- !u!114 &526344201
@@ -6361,6 +6408,7 @@ RectTransform:
   - {fileID: 1458278302}
   - {fileID: 1254632427}
   - {fileID: 1462654761}
+  - {fileID: 301490248}
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -6823,7 +6871,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 1969722462}
   m_Direction: 2
   m_Value: 0
-  m_Size: 0.9686385
+  m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -7984,7 +8032,7 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls: []
   m_Sprite: {fileID: 0}
-  m_Type: 0
+  m_Type: 1
   m_PreserveAspect: 0
   m_FillCenter: 1
   m_FillMethod: 4
@@ -11166,8 +11214,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 526344200}
-        m_TargetAssemblyTypeName: PageContainer, Assembly-CSharp
+      - m_Target: {fileID: 2100211433}
+        m_TargetAssemblyTypeName: SaveStoryButton, Assembly-CSharp
         m_MethodName: CloseSubmissionStatus
         m_Mode: 1
         m_Arguments:
@@ -12663,6 +12711,7 @@ GameObject:
   - component: {fileID: 2100211436}
   - component: {fileID: 2100211435}
   - component: {fileID: 2100211434}
+  - component: {fileID: 2100211433}
   m_Layer: 5
   m_Name: SaveStoryButton
   m_TagString: Untagged
@@ -12690,6 +12739,21 @@ RectTransform:
   m_AnchoredPosition: {x: -102.75, y: 56}
   m_SizeDelta: {x: 48, y: 48}
   m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &2100211433
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2100211431}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe085da07e3ebcc46b50c5f7f7a4ce13, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  saveStoryHandler: {fileID: 301490247}
+  StoryNamePrompt: {fileID: 1254632426}
+  StorySubmissionStatus: {fileID: 1462654760}
 --- !u!114 &2100211434
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -12734,8 +12798,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 526344200}
-        m_TargetAssemblyTypeName: PageContainer, Assembly-CSharp
+      - m_Target: {fileID: 2100211433}
+        m_TargetAssemblyTypeName: SaveStoryButton, Assembly-CSharp
         m_MethodName: OpenStoryNameMenu
         m_Mode: 1
         m_Arguments:
