+++ Assets/Scenes/Story Builder/Page Container/PageContainer.cs
@@ -27,8 +27,6 @@ public class PageContainer : MonoBehaviour
 
     [Header("Scene References")]
     public GameObject pageIconContainer;
-    public GameObject StoryNamePrompt;
-    public GameObject StorySubmissionStatus;
 
     //
     [Header("Settings")]
@@ -305,34 +303,4 @@ public class PageContainer : MonoBehaviour
 
     }
 
-    private void PutStoryInDatabase(string storyName)
-    {
-        List<StoryPage> storyPages = GetPagesForStorySubmission();
-        Story story = new Story(storyPages); // story with no name or font
-        story.learnerId = LearnerLogin.staticLearner._id; // should make another constructor if these get made more often
-        story.storyName = storyName;
-        StartCoroutine(ServerRequestHandler.PostStory(story));
-    }
-
-    public void OpenStoryNameMenu()
-    {
-        StoryNamePrompt.SetActive(true);
-    }
-    public void CloseStoryNameMenu()
-    {
-        string storyName = StoryNamePrompt.GetComponentInChildren<Text>().text;
-        StoryNamePrompt.SetActive(false);
-        PutStoryInDatabase(storyName);
-        // if story submitted successfully... (should check this eventually)
-        DisplaySubmissionStatus();
-    }
-    private void DisplaySubmissionStatus()
-    {
-        // might be nice to have the success message contain the name of the submitted story
-        StorySubmissionStatus.SetActive(true);
-    }
-    public void CloseSubmissionStatus()
-    {
-        StorySubmissionStatus.SetActive(false);
-    }
 }
