+++ Assets/Scenes/Story Builder/Save Story Button/SaveStoryButton.cs
@@ -5,9 +5,10 @@ using UnityEngine.UI;
 
 public class SaveStoryButton : MonoBehaviour
 {
-    [Header("Page Icon Containers")]
-    public PageContainer pageContainer;
     private Button button;
+    public GameObject saveStoryHandler;
+    public GameObject StoryNamePrompt;
+    public GameObject StorySubmissionStatus;
 
     private void Start()
     {
@@ -15,25 +16,26 @@ public class SaveStoryButton : MonoBehaviour
         //button.onClick.AddListener(SaveStory);
     }
 
-    // no real story saving implemented just yet
-    // private void SaveStory()
-    // {
-    //     List<StoryPage> toSave = pageContainer.GetAllSentencesInPages();
+    public void OpenStoryNameMenu()
+    {
+        StoryNamePrompt.SetActive(true);
+    }
+    public void CloseStoryNameMenu()
+    {
+        string storyName = StoryNamePrompt.GetComponentInChildren<Text>().text;
+        StoryNamePrompt.SetActive(false);
+        saveStoryHandler.GetComponent<SaveStoryHandler>().PutStoryInDatabase(storyName);
+        // if story submitted successfully... (should check this eventually)
+        DisplaySubmissionStatus();
+    }
+    private void DisplaySubmissionStatus()
+    {
+        // might be nice to have the success message contain the name of the submitted story
+        StorySubmissionStatus.SetActive(true);
+    }
+    public void CloseSubmissionStatus()
+    {
+        StorySubmissionStatus.SetActive(false);
+    }
 
-    //     SaveStoryHandler.SaveStory(toSave);
-
-    //     // for(int i = 0; i < toSave.Count; i++)
-    //     // {
-    //     //     string blah = "";
-
-    //     //     for (int o = 0; o < toSave[i].wordTiles.Count; o++)
-    //     //     {
-                
-
-    //     //         blah += toSave[i].wordTiles[o].textToDisplay;
-    //     //     }
-
-    //     //     Debug.Log(blah);
-    //     // }
-    // }
 }
